requirements.txt
autopep8==1.5.3
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cssselect2==0.3.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2
django-allauth==0.41.0
django-cors-headers==3.2.1
django-extra-fields==2.0.5
django-filter==2.2.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.11.0
gunicorn==20.0.4
html5lib==1.0.1
idna==2.9
itypes==1.1.0
Jinja2==2.11.1
Markdown==3.2.1
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
Pillow==7.1.1
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pycparser==2.20
Pyphen==0.9.5
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
sentry-sdk==0.14.3
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
tinycss2==1.0.2
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.8
WeasyPrint==51
webencodings==0.5.1
whitenoise==3.3.1
